Moved formatTime function in communfuntions file
